+++ src/deviceatw.js
@@ -806,6 +806,9 @@ class DeviceAtw extends EventEmitter {
                         this.emit('error', error);
                     };
                 })
+                .on('success', (message) => {
+                    this.emit('success', message);
+                })
                 .on('message', (message) => {
                     this.emit('message', message);
                 })