+++ src/deviceerv.js
@@ -559,6 +559,9 @@ class DeviceErv extends EventEmitter {
                         this.emit('error', error);
                     };
                 })
+                .on('success', (message) => {
+                    this.emit('success', message);
+                })
                 .on('message', (message) => {
                     this.emit('message', message);
                 })