+++ src/melcloud.js
@@ -36,7 +36,9 @@ class MelCloud extends EventEmitter {
                 } catch (error) {
                     this.emit('error', `Impulse generator error: ${error}.`);
                 };
-            }).on('state', (state) => { });
+            }).on('state', (state) => {
+                const emitState = state ? this.emit('success', `Impulse generator started.`) : this.emit('warn', `Impulse generator stopped.`);
+            });
         };
     };