+++ src/melclouderv.js
@@ -37,7 +37,9 @@ class MelCloudErv extends EventEmitter {
             } catch (error) {
                 this.emit('error', `Impulse generator error: ${error}.`);
             };
-        }).on('state', (state) => { });
+        }).on('state', (state) => {
+            const emitState = state ? this.emit('success', `Impulse generator started.`) : this.emit('warn', `Impulse generator stopped.`);
+        });
     };
 
     async checkState() {